refactor callback functions to make code easier to read and modifiy routes/index.js file
